Add test for none compression
